refactor: Unsubscriber, 配信停止系のメソッドをUnsubscriberモデルにまとめる
